perf(admin): async-chunk Tiptap for About CMS editor (INF-106)
Move the Tiptap implementation to rich-text-editor-core and load it from
rich-text-editor via next/dynamic with ssr:false (client component only).
Keeps a toolbar-shaped loading skeleton. Drops ~120kB from /admin/about first
load JS per next build; Tiptap is no longer in the admin/about page chunk.
